+++ _black_hole_spec/src/lib.rs
@@ -138,11 +138,11 @@ pub struct QuarkModelConfig {
     ///
     /// When set, quark flips `frozen` <-> `unfrozen` every `oscillation_steps`
     /// completed optimize calls after any configured offset.
-    pub oscillation_steps: Option<usize>,
+    pub oscillation_steps: Option<u32>,
     /// Optional number of optimize steps to wait before oscillation begins.
     ///
     /// Ignored when `oscillation_steps` is `None`.
-    pub oscillation_offset: Option<usize>,
+    pub oscillation_offset: Option<u32>,
     /// Optional checkpoint object to load model weights from for this instance.
     ///
     /// When `None`, quark loads weights from its configured server model path.
